further refactorings
* dropped commit preprocessing (we now have a factory function that make datamodel.Commit from a git.Commit,
aka RawCommit
* old feature extraction functions are now rule_helpers
