+++ prospector/datamodel/constants.py
@@ -28,12 +28,3 @@ RELEVANT_EXTENSIONS = [
     "yml",
     "jar",
 ]
-
-ALLOWED_SITES = [
-    "for.testing.purposes",
-    "lists.apache.org",
-    "just.an.example.site",
-    "one.more.example.site",
-    "non-existing-url.com",  # for testing.
-    "jvndb.jvn.jp",  # for trying out: usually does not aviable, but not always, anyway it is a good example
-]
